requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langgraph>=0.0.20
langchain-openai>=0.0.2
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

